Using longs to represent sizes in bytes

+++ dht/src/main/java/sustain/synopsis/dht/store/AllocationPolicy.java
@@ -6,5 +6,5 @@ import java.util.List;
  * Defines a policy for allocating available storage directories to continuous storage requests.
  */
 public interface AllocationPolicy {
-    DiskManager.StorageDirectory select(int requestedCapacity, List<DiskManager.StorageDirectory> dirs);
+    DiskManager.StorageDirectory select(long requestedCapacity, List<DiskManager.StorageDirectory> dirs);
 }
